+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerExceptionMapper.kt
@@ -2,17 +2,17 @@ package com.bitmovin.analytics.exoplayer
 
 import com.bitmovin.analytics.data.ErrorCode
 import com.bitmovin.analytics.data.LegacyErrorData
-import com.bitmovin.analytics.error.ExceptionMapper
 import com.bitmovin.analytics.features.errordetails.ErrorData
 import com.bitmovin.analytics.utils.topOfStacktrace
 import com.google.android.exoplayer2.ExoPlaybackException
+import com.google.android.exoplayer2.PlaybackException
 import com.google.android.exoplayer2.source.BehindLiveWindowException
 import com.google.android.exoplayer2.upstream.HttpDataSource
 
-internal class ExoPlayerExceptionMapper : ExceptionMapper<Throwable> {
+internal object ExoPlayerExceptionMapper {
 
-    private val errorMessages = mapOf(
-        -1 to "Unknown Error",
+    private val errorTypeMap = mapOf(
+        -1 to "Unknown Error Type",
         0 to "Source Error",
         1 to "Render Error",
         2 to "Unexpected Error",
@@ -20,41 +20,38 @@ internal class ExoPlayerExceptionMapper : ExceptionMapper<Throwable> {
         4 to "Out of memory Error",
     )
 
-    override fun map(throwable: Throwable): ErrorCode {
-        val exceptionType = getExceptionType(throwable)
-        val exceptionMessage = throwable.message ?: ""
-        val exceptionDetails = throwable.topOfStacktrace
-        val prefix = errorMessages[exceptionType] ?: "Unknown Error"
-
+    fun map(playbackException: PlaybackException): ErrorCode {
         val errorMessage: String
-        val errorCodeDescription: String
+        val cause = playbackException.cause
+        val exceptionMessage = playbackException.message ?: ""
 
-        when (val exception = throwable.cause ?: throwable) {
-            is HttpDataSource.InvalidResponseCodeException -> {
-                errorCodeDescription = "$prefix: InvalidResponseCodeException (Status Code: ${exception.responseCode}, URI: ${exception.dataSpec.uri})"
-                errorMessage = "Data Source request failed with HTTP status: ${exception.responseCode} - ${exception.dataSpec.uri}"
+        if (cause != null) {
+            when (cause) {
+                is HttpDataSource.InvalidResponseCodeException -> {
+                    errorMessage = "Data Source request failed with HTTP status: ${cause.responseCode} - ${cause.dataSpec.uri}"
+                }
+                is HttpDataSource.InvalidContentTypeException -> {
+                    errorMessage = "Invalid Content Type: ${cause.contentType}"
+                }
+                is HttpDataSource.HttpDataSourceException -> {
+                    errorMessage = "Unable to connect: ${cause.dataSpec.uri}"
+                }
+                is BehindLiveWindowException -> {
+                    errorMessage = "Behind live window: required segments not available"
+                }
+                else -> {
+                    errorMessage = "$exceptionMessage - $cause"
+                }
             }
-            is HttpDataSource.InvalidContentTypeException -> {
-                errorCodeDescription = "$prefix: InvalidContentTypeException (ContentType: ${exception.contentType})"
-                errorMessage = "Invalid Content Type: ${exception.contentType}"
-            }
-            is HttpDataSource.HttpDataSourceException -> {
-                errorCodeDescription = "$prefix: HttpDataSourceException (URI: ${exception.dataSpec.uri})"
-                errorMessage = "Unable to connect: ${exception.dataSpec.uri}"
-            }
-            is BehindLiveWindowException -> {
-                errorCodeDescription = "$prefix: BehindLiveWindowException"
-                errorMessage = "Behind live window: required segments not available"
-            }
-            else -> {
-                errorCodeDescription = "$prefix: $exceptionMessage"
-                errorMessage = exceptionMessage
-            }
+        } else {
+            errorMessage = exceptionMessage
         }
-        val legacyErrorData = LegacyErrorData(errorMessage, exceptionDetails)
 
-        val errorData = ErrorData(errorMessage, exceptionDetails.toList())
-        return ErrorCode(exceptionType, errorCodeDescription, errorData, legacyErrorData)
+        val topOfStackTrace = playbackException.topOfStacktrace
+        val legacyErrorData = LegacyErrorData(errorMessage, topOfStackTrace)
+        val errorData = ErrorData(errorMessage, topOfStackTrace.toList())
+        val errorCodeDescription = errorTypeMap[getExceptionType(playbackException)] + ": " + playbackException.errorCodeName
+        return ErrorCode(playbackException.errorCode, errorCodeDescription, errorData, legacyErrorData)
     }
 
     private fun getExceptionType(throwable: Throwable): Int {